Importing the pt_BR work by Vinícius Coque <vcoque@gmail.com>
IMPORTANT: this branch should not go into the tree unless one of the orignal authors submit it.

This has approx. 79% of the BR_PT strings in it...  Here's how I made it,

I took the old work, which was buried in the earliest OPNsense fork branch,
  $ git checkout -b opn_firstcommit ff4b1affcdb881b809056f1b77413a03a8c61cd0
  $ cp ./usr/local/share/locale/pt_BR.ISO8859-1/LC_MESSAGES/pfSense.po \
       /tmp/pfSense.pt_BR.ISO8859-1.po
  $ git checkout pt_br_bootstrapping

Then I merged the file in using gettext tools,
  $ cd lang/
  $ msgmerge --multi-domain --add-location --previous \
      -o pt_BR.po /tmp/pfSense.pt_BR.ISO8859-1.po en_US.pot

From there, edit the headers a bit to connect it to the OPNsense project, and the work is salvaged!

Signed-off-by: Isaac (.ike) Levy <ike@blackskyresearch.net>